Made the Gpa only display three decimals

+++ app/src/main/java/com/example/mr_goaty_/gpacalculator/MainActivity.java
@@ -33,7 +33,7 @@ public class MainActivity extends AppCompatActivity {
         button.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                int gpa = model.calculateGPA();
+                double gpa = model.calculateGPA();
                 final TextView text = findViewById(R.id.GPA_Value);
                 text.setText("" + gpa + "");
             }
